refactor(ui): share bounded image preview resources
