+++ crates/ui/src/lib.rs
@@ -28,6 +28,7 @@ pub mod files;
 pub mod frost;
 pub mod history;
 pub mod icons;
+pub(crate) mod image_media;
 pub mod links;
 pub mod loaders;
 pub mod markdown;
